+++ packages/react-core/src/index.d.ts
@@ -12,6 +12,7 @@ export { InvalidColumnError } from './utils/InvalidColumnError';
 export { debounce } from './utils/debounce';
 export { throttle } from './utils/throttle';
 export { randomString } from './utils/randomString';
+export { assert as _assert } from './utils/assert';
 
 export { makeIntervalComplete } from './utils/makeIntervalComplete';