+++ packages/react-widgets/src/types.d.ts
@@ -98,6 +98,7 @@ export type LegendWidget = {
   className?: string;
   initialCollapsed?: boolean;
   customLegendTypes?: Record<string, Function>;
+  layerOrder?: string[];
 }
 
 export type WidgetWithAlert = {